refactor(api): define the public types in camas.v0; the engine consumes them
Per review: the namespace was inverted — v0 re-exported definitions
living in camas.core, so the contract file was a facade and the API
history was scattered. Now the public types are defined in
camas/v0/{task,completion,task_event,leaf_state,effect}.py and
camas.core/camas.main consume them, never the reverse: the diff history
of src/camas/v0/ is the complete history of the public API, and
Task.__module__ et al. point at the public address.

Engine-side internals stay unversioned in core (task_label,
VarBinding/MatrixBinding, ChainLink/LeafInfo/next_state, EventSink,
TaskResult/RunResult); core/task_event.py moves wholesale (it was all
public). Group joins the contract — it is the documented isinstance
target for "either grouping kind" and the base of two contract types.

tests/test_v0.py now also asserts every contract class is defined under
camas.v0.*, so the inversion itself is pinned.
