+++ README.md
@@ -139,7 +139,7 @@ Define an `Effect` in your `tasks.py` and it's discovered automatically — usab
 
 ## Versioning
 
-The public API lives behind [`camas.v0`](https://github.com/JPHutchins/camas/blob/main/src/camas/v0/__init__.py), versioned the way semver versions the package: `v0` pairs with camas 0.x and is exactly as loose as semver says 0.x is. The surface prefers to grow — new names, fields appended with defaults — but breaking changes remain possible until 1.0, made deliberately and noted in releases. At 1.0 the contract hardens: a stable-era namespace never removes or changes an exported name, a breaking change forces the next `camas.vN`, and old namespaces keep shipping, so a `tasks.py` or effect plugin written against one keeps working across upgrades.
+The public API is defined in [`camas.v0`](https://github.com/JPHutchins/camas/tree/main/src/camas/v0) — the internals consume it, never the reverse, so the diff history of `src/camas/v0/` is the complete history of the API. It is versioned the way semver versions the package: `v0` pairs with camas 0.x and is exactly as loose as semver says 0.x is. The surface prefers to grow — new names, fields appended with defaults — but breaking changes remain possible until 1.0, made deliberately and noted in releases. At 1.0 the contract hardens: a stable-era namespace never removes or changes an exported name, a breaking change forces the next `camas.vN`, and old namespaces keep shipping, so a `tasks.py` or effect plugin written against one keeps working across upgrades.
 
 The top-level surface (`from camas import Task, Sequential, Parallel, Effect`) re-exports the definers from the latest version namespace — best effort across major generations, fine for a `tasks.py` that lives next to its dev environment. The full plugin contract — `TaskNode`, the `TaskEvent` stream, `LeafState`, `Completion` — lives only in `camas.v0`. Everything else (`camas.core.*`, `camas.main.*`) is internal: importable today, but with no stability promise.
