+++ src/camas/core/completion.py
@@ -1,48 +1,20 @@
 # SPDX-License-Identifier: MIT
 # SPDX-FileCopyrightText: 2026 JP Hutchins
 
-"""Completion outcomes (``Finished`` | ``Skipped``) and the per-run result aggregates."""
+"""Per-run result aggregates over the :mod:`camas.v0` completion outcomes."""
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, NamedTuple, TypeAlias
+from typing import TYPE_CHECKING, NamedTuple
 
 if TYPE_CHECKING:
-	from collections.abc import Sequence
-
-
-class Finished(NamedTuple):
-	"""Completion outcome: task ran to exit with a returncode.
-
-	>>> Finished(0, 1.234, (b"all clean",))
-	Finished(returncode=0, elapsed=1.234, output=(b'all clean',))
-	"""
-
-	returncode: int
-	elapsed: float
-	output: Sequence[bytes]
-
-
-class Skipped(NamedTuple):
-	"""Completion outcome: task was skipped due to a prior Sequential failure.
-
-	Carries the returncode of the task that caused the skip — an Either-like
-	propagation so callers that need an rc (e.g. the overall run's exit code)
-	can read it uniformly across completion variants.
-
-	>>> Skipped(1)
-	Skipped(returncode=1)
-	"""
-
-	returncode: int
-
-
-Completion: TypeAlias = Finished | Skipped
+	from ..v0 import Completion
 
 
 class TaskResult(NamedTuple):
 	"""Result of a single completed task.
 
+	>>> from camas.v0 import Finished
 	>>> TaskResult("lint", Finished(0, 1.234, (b"all clean",)))
 	TaskResult(name='lint', completion=Finished(returncode=0, elapsed=1.234, output=(b'all clean',)))
 	"""
@@ -54,6 +26,7 @@ class TaskResult(NamedTuple):
 class RunResult(NamedTuple):
 	"""Result of running an entire task tree.
 
+	>>> from camas.v0 import Finished
 	>>> RunResult(0, (TaskResult("a", Finished(0, 0.1, ())),), 0.1)
 	RunResult(returncode=0, results=(TaskResult(name='a', completion=Finished(returncode=0, elapsed=0.1, output=())),), elapsed=0.1)
 	"""
