+++ src/camas/core/matrix.py
@@ -17,7 +17,8 @@ if sys.version_info >= (3, 11):
 else:  # pragma: no cover
 	from typing_extensions import assert_never
 
-from .task import MatrixBinding, Parallel, Sequential, Task, TaskNode, VarBinding, task_label
+from ..v0 import Parallel, Sequential, Task, TaskNode
+from .task import MatrixBinding, VarBinding, task_label
 
 if TYPE_CHECKING:
 	from collections.abc import Mapping
