+++ src/camas/core/render.py
@@ -15,9 +15,10 @@ if sys.version_info >= (3, 11):
 else:  # pragma: no cover
 	from typing_extensions import assert_never
 
+from ..v0 import Parallel, Sequential, Task, TaskNode
 from .leaf_state import ChainLink, LeafInfo
 from .matrix import expand_matrix
-from .task import Parallel, Sequential, Task, TaskNode, task_label
+from .task import task_label
 
 if TYPE_CHECKING:
 	from collections.abc import Iterator, Mapping
