+++ src/camas/core/traversal.py
@@ -13,8 +13,8 @@ if sys.version_info >= (3, 11):
 else:  # pragma: no cover
 	from typing_extensions import assert_never
 
+from ..v0 import Parallel, Sequential, Task, TaskNode
 from .leaf_state import ChainLink, LeafInfo
-from .task import Parallel, Sequential, Task, TaskNode
 
 if TYPE_CHECKING:
 	from collections.abc import Iterator
