+++ src/camas/main/state.py
@@ -21,8 +21,7 @@ if TYPE_CHECKING:
 	from collections.abc import Mapping
 	from pathlib import Path
 
-	from ..core.effect import Effect
-	from ..core.task import TaskNode
+	from ..v0 import Effect, TaskNode
 
 
 class LoadOk(NamedTuple):
