+++ tests/core/test_effect.py
@@ -15,15 +15,13 @@ else:  # pragma: no cover
 	from exceptiongroup import BaseExceptionGroup
 
 from camas import Parallel, Sequential, Task
-from camas.core.completion import Skipped
 from camas.core.execution import run
-from camas.core.leaf_state import Completed, LeafState
-from camas.core.task_event import CompletedEvent, StartedEvent, TaskEvent
+from camas.v0 import Completed, CompletedEvent, LeafState, Skipped, StartedEvent, TaskEvent
 
 if TYPE_CHECKING:
 	from collections.abc import Sequence
 
-	from camas.core.task import TaskNode
+	from camas.v0 import TaskNode
 
 
 class RecorderCtx(NamedTuple):
