+++ tests/core/test_execution.py
@@ -9,8 +9,8 @@ import time
 import pytest
 
 from camas import Parallel, Sequential, Task
-from camas.core.completion import Finished
 from camas.core.execution import run
+from camas.v0 import Finished
 
 
 def test_force_color_injected_in_subprocess(monkeypatch: pytest.MonkeyPatch) -> None:
