+++ tests/effect/test_status.py
@@ -10,9 +10,7 @@ from typing import TYPE_CHECKING, TypeVar
 import pytest
 
 from camas import Parallel, Sequential, Task
-from camas.core.completion import Finished, Skipped
-from camas.core.leaf_state import LeafState, Waiting, next_state
-from camas.core.task_event import CompletedEvent, OutputEvent, StartedEvent, TaskEvent
+from camas.core.leaf_state import next_state
 from camas.core.traversal import flatten_leaves
 from camas.effect.status import (
 	Active,
@@ -21,9 +19,19 @@ from camas.effect.status import (
 	Status,
 	StatusOptions,
 )
+from camas.v0 import (
+	CompletedEvent,
+	Finished,
+	LeafState,
+	OutputEvent,
+	Skipped,
+	StartedEvent,
+	TaskEvent,
+	Waiting,
+)
 
 if TYPE_CHECKING:
-	from camas.core.effect import Effect
+	from camas.v0 import Effect
 
 TS = datetime(2026, 5, 21, 14, 30, 0, 750_000)
