+++ tests/effect/test_summary.py
@@ -8,17 +8,23 @@ from datetime import datetime
 from typing import TYPE_CHECKING, TypeVar
 
 from camas import Parallel, Sequential, Task
-from camas.core.completion import Finished, Skipped
-from camas.core.leaf_state import LeafState, Waiting
-from camas.core.task_event import CompletedEvent, StartedEvent, TaskEvent
 from camas.effect.summary import Fixed, Summary, SummaryOptions
+from camas.v0 import (
+	CompletedEvent,
+	Finished,
+	LeafState,
+	Skipped,
+	StartedEvent,
+	TaskEvent,
+	Waiting,
+)
 
 if TYPE_CHECKING:
 	from collections.abc import Callable
 
 	import pytest
 
-	from camas.core.effect import Effect
+	from camas.v0 import Effect
 
 TS = datetime(2026, 5, 21, 14, 30, 0)
